Code rework: move evm APIs inside the evm family
